Service-based_IL/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

joblib==1.5.1
matplotlib==3.9.2
numpy==1.26.1
pyarrow==19.0.0
pandas==2.3.2
pillow==11.1.0
plotly==6.5.0
psutil==7.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pyzmq==27.0.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
streamlit==1.48.1
streamlit-option-menu==0.4.0
torch==2.5.1
xgboost==2.1.3
